reorganised compile_metadata to keep only top level functions. Second level functions were moved to hydra_helper. Also, get_imgstore_name will return imgstore_names in mac format (compatible to pathlib in any platform).
