+++ tierpsytools/hydra/compile_metadata.py
@@ -196,7 +196,6 @@ def merge_robot_metadata(
     return robot_metadata
 
 #%%
-
 def populate_96WPs(worm_sorter,
                    n_columns=12,
                    n_rows=8,
@@ -300,43 +299,12 @@ def populate_96WPs(worm_sorter,
         with open(wormsorter_log_fname, 'a') as fid:
             fid.write ('No wormsorter plate errors \n')
 
-    cols_to_keep = ['hydra_number', 'imaging_plate_id', 'imaging_run',
-                    'well_name', 'worm_strain', 'worm_code', 'worm_gene',
-                    'bacteria_strain', 'start_well','end_well',
-                    'cave_humidity_percent', 'cave_temp_oC', 'cave_time',
-                    'comments', 'date_bleached_yyyymmdd', 'days_in_diapause',
-                    'media_type', 'date_plates_poured_yyyymmdd',
-                    'date_refed_yyyymmdd','number_worms_per_well',
-                    'recording_time']
-    plate_metadata = plate_metadata.reindex(columns=cols_to_keep)
     plate_metadata.to_csv(saveto, index=False)
 
     return plate_metadata
 
-# %% separate processing of bad wells
-
-def convert_bad_wells_lut(bad_wells_csv):
-    """
-    author: @ilbarlow
-    Function for converting bad_wells_csv for input into dataframes
-    Input:
-    bad_wells_csv -.csv file listing imaging_plate_id and well_name
-
-    Output:
-    bad_wells_df - DataFrame with columns imaging_plate_id, well_name and
-    is_bad_well=True
-    """
-
-    bad_wells_df = pd.read_csv(bad_wells_csv)
-                                         # encoding='utf-8-sig')
-    bad_wells_df['is_bad_well'] = True
-
-    return bad_wells_df
-
 
 # %%extra function if using both robot and wormsorter
-
-
 def merge_robot_wormsorter(day_root_dir,
                            robot_metadata,
                            plate_metadata,
@@ -357,6 +325,7 @@ def merge_robot_wormsorter(day_root_dir,
     Ouput:
     concat_metadata - dataframe that can be used in get_day_metadata
     """
+    from tierpsytools.hydra.hydra_helper import convert_bad_wells_lut
 
     # saving and overwriting checks
     if saveto is None:
@@ -495,178 +464,6 @@ def get_camera_serial(
     return out_metadata
 
 
-def add_imgstore_name(
-        metadata, raw_day_dir, n_wells=96, run_number_regex=r'run\d+_'
-        ):
-    """
-    @author: em812
-    Add the imgstore name of the hydra videos to the day metadata dataframe.
-
-    param:
-        metadata = pandas dataframe
-            Dataframe with metadata for a given day of experiments.
-            See README.md for details on fields.
-        raw_day_dir = path to directory
-            RawVideos root directory of the specific day, where the
-            imgstore names can be found.
-        n_wells = integer
-            Number of wells in imaging plate (only 96 supported at the
-            moment)
-
-    return:
-        out_metadata = metadata dataframe with imgstore_name added
-
-    """
-    from os.path import join
-    from tierpsytools.hydra.hydra_helper import run_number_from_regex
-
-    ## Checks
-    # - check if raw_day_dir exists
-    if not raw_day_dir.exists:
-        warnings.warn("\nRawVideos day directory was not found. "
-                      +"Imgstore names cannot be added to the metadata.\n",
-                      +"Path {} not found.".format(raw_day_dir))
-        return metadata
-
-    # - if the raw_dat_dir contains a date in yyyymmdd format, check if the
-    #   date in raw_day_dir matches the date of runs stored in the metadata
-    #   dataframe
-    date_of_runs = metadata['date_yyyymmdd'].astype(str).values[0]
-    date_in_dir = re.findall(r'(20\d{6})',raw_day_dir.stem)
-    if len(date_in_dir)==1 and date_of_runs != date_in_dir[0]:
-        warnings.warn(
-            '\nThe date in the RawVideos day directory does not match ' +
-            'the date_yyyymmdd in the day metadata dataframe. '
-            'Imgstore names cannot be added to the metadata.\n'+
-            'Please check the dates and try again.')
-        return metadata
-
-    # add camera serial number to metadata
-    metadata = get_camera_serial(metadata, n_wells=n_wells)
-
-    # get imgstore full paths = raw video directories that contain a
-    # metadata.yaml file and get the run and camera number from the names
-    file_list = [file for file in raw_day_dir.rglob("metadata.yaml")]
-    #print('There are {} raw videos found in {}.\n'.format(
-    #    len(file_list),raw_day_dir))
-    camera_serial = [str(file.parent.parts[-1]).split('.')[-1]
-                               for file in file_list]
-
-    imaging_run_number = run_number_from_regex(file_list, run_number_regex)
-    # imaging_run_number = run_number_from_timestamp(file_list, camera_serial)
-
-    file_meta = pd.DataFrame({
-        'file_name': file_list,
-        'camera_serial': camera_serial,
-        'imaging_run_number': imaging_run_number
-        })
-
-    # keep only short imgstore_name (experiment_day_dir/imgstore_name_dir)
-    file_meta['imgstore_name'] = file_meta['file_name'].apply(
-            lambda x: join(*x.parts[-3:-1]))
-
-    # merge dataframes to store imgstore_name for each metadata row
-    out_metadata = pd.merge(
-            metadata,
-            file_meta[['imaging_run_number','camera_serial','imgstore_name']],
-            how='outer',on=['imaging_run_number','camera_serial'])
-
-    ## Checks
-    # - check if there are multiple videos with the same day, run and camera
-    #   number. If yes, raise a warning (it is not necessarily an error, but
-    #   good to let the user know).
-    # if the imgstore names are more than 3 times the unique combinations of
-    # camera serial and run number, then there are videos that have the same run
-    # number are camera serial by mistake.
-    n_not_unique = file_meta.shape[0]- \
-        3*file_meta.drop_duplicates(subset=['camera_serial', 'imaging_run_number']).shape[0]
-    if n_not_unique!=0:
-        warnings.warn('\n\nThere are {} sets of '.format(n_not_unique)
-                      +'videos with the same day, run and camera number.\n\n'
-                     )
-    # - check if there are missing videos (we expect to have videos from every
-    #   camera of a given instrument). If yes, raise a warning.
-    breakpoint()
-    if out_metadata['imgstore_name'].isna().sum()>0:
-        not_found = out_metadata.loc[out_metadata['imgstore_name'].isna(),
-                                 ['imaging_run_number', 'camera_serial']]
-        for i,row in not_found.iterrows():
-            warnings.warn('\n\nNo video found for day '
-                          +'{}, run {}, camera {}.\n\n'.format(
-                                  raw_day_dir.stem,*row.values)
-                          )
-
-    return out_metadata
-
-
-def get_date_of_runs_from_aux_files(manual_metadata_file):
-    """
-    @author: em812
-    Finds the date of the runs in the manual_metadata_file.
-    If the date field is missing, then it looks for the date of runs in the
-    manual_metadata_file file name (in the format yyyymmdd).
-    If there is no date in this format in the file name, then it looks at the
-    folder name (which is the folder for a specific day of experiments).
-    If the date in yyyymmdd format cannot be found in any of these locations,
-    an error is raised.
-
-    param:
-        manual_metadata_file: full path to .csv file
-            Full path to the manual metadata file
-
-    return:
-        date_of_runs: string
-            The date of the experiments, in yyyymmdd format
-    """
-    manual_metadata = pd.read_csv(manual_metadata_file, index_col=False)
-    if 'date_yyyymmdd' in manual_metadata.columns:
-        date_of_runs = manual_metadata['date_yyyymmdd'].astype(str).values[0]
-    else:
-        date_of_runs = re.findall(r'(20\d{6})',manual_metadata_file.stem)
-        if len(date_of_runs)==1:
-            date_of_runs = date_of_runs[0]
-        else:
-            date_of_runs = re.findall(r'(20\d{6})',
-                                      manual_metadata_file.parent.stem)
-            if len(date_of_runs)==1:
-                date_of_runs = date_of_runs[0]
-            else:
-                raise ValueError('The date of the experiments cannot be '
-                                 +'identified in the auxiliary files path '
-                                 +'names. Please add a data_yyyymmdd column '
-                                 +'to the manual_metadata file.'
-                                 )
-
-    # If the aux_day_dir contains the date, then make sure it matches the date
-    # extracted from the manual_metadate_file
-    date_in_dir = re.findall(r'(20\d{6})',manual_metadata_file.parent.stem)
-    if len(date_in_dir)==1 and date_in_dir[0]!=date_of_runs:
-        raise ValueError('\nThe date_of_runs taken from the '
-                         +'manual_metadata_file ({}) '.format(date_of_runs)
-                         +'does not match the date of runs in the folder '
-                         +'name {}.\n'.format(manual_metadata_file.parent)
-                         +'Please set the correct date and try again.')
-    return date_of_runs
-
-def check_dates_in_yaml(metadata,raw_day_dir):
-    """
-    Checks the day metadata, to make sure that the experiment date stored in
-    the metadata dataframe matches the date in the metadata.yaml file in the
-    corresponding raw video directory.
-
-    param:
-        metadata : pandas dataframe
-            Dataframe containing all the metadata from one day of experiments
-        raw_day_dir : directory path
-            Path of the directory containing the RawVideos for the
-            specific day of experiments.
-
-    return:
-        None
-    """
-
-    return
-
 def get_day_metadata(
         imaging_plate_metadata, manual_metadata_file,
         merge_on=['imaging_plate_id'], n_wells=96,
@@ -719,6 +516,10 @@ def get_day_metadata(
         metadata: pandas dataframe
             dataframe with day metadata
     """
+    from tierpsytools.hydra.hydra_helper import get_date_of_runs_from_aux_files
+
+    manual_metadata_file = Path(manual_metadata_file)
+
     #find the date of the hydra experiments
     date_of_runs = get_date_of_runs_from_aux_files(manual_metadata_file)
 
@@ -768,7 +569,7 @@ def get_day_metadata(
                     )
         if raw_day_dir.exists:
             metadata = add_imgstore_name(
-                    metadata,raw_day_dir,n_wells=n_wells,
+                    metadata, raw_day_dir, n_wells=n_wells,
                     run_number_regex=run_number_regex
                     )
             #check_dates_in_yaml(metadata,raw_day_dir)
@@ -785,8 +586,6 @@ def get_day_metadata(
     return metadata
 
 # %% Check day metadata is the correct side
-
-
 def day_metadata_check(day_metadata, day_root_dir, plate_size=96):
 
     """@author: ibarlow
@@ -940,173 +739,4 @@ def concatenate_days_metadata(
 
 #%%
 if __name__ == '__main__':
-    # Example 1:
-    # Input
-    day_root_dir = Path('/Users/em812/Data/Hydra_pilot/AuxiliaryFiles/'
-                        +'20191108_tierpsytools_dev')
-    raw_day_dir = Path('/Volumes/behavgenom$/Ida/Data/Hydra/PilotDrugExps/'
-                        +'RawVideos/20191108')
-    sourceplate_file = day_root_dir / '20191107_sourceplates.csv'
-    manual_meta_file = day_root_dir / '20191108_manual_metadata.csv'
-
-    # Save to
-    #robot_metadata_file = day_root_dir.joinpath('20191107_robot_metadata.csv')
-    metadata_file = day_root_dir.joinpath('20191108_day_metadata.csv')
-
-    # Run
-    robot_metadata = merge_robot_metadata(sourceplate_file, saveto=False)
-    day_metadata = get_day_metadata(robot_metadata, manual_meta_file,
-                                    saveto=metadata_file,del_if_exists=True,
-                                    raw_day_dir=raw_day_dir)
-
-    # Example 2:
-    # Input
-    aux_root_dir = Path(
-            '/Volumes/behavgenom$/Ida/Data/Hydra/PilotDrugExps/AuxiliaryFiles')
-    day_root_dirs = [d for d in aux_root_dir.glob("*") if d.is_dir()]
-
-    sourceplate_files = [
-            [file for file in d.glob('*_sourceplates.csv')]
-            for d in day_root_dirs
-            ]
-    manual_meta_files = [
-            [file for file in d.glob('*_manual_metadata.csv')]
-            for d in day_root_dirs
-            ]
-
-    # Saveto
-    metadata_files = [d.joinpath('{}_day_metadata.csv'.format(d.stem))
-            for d in day_root_dirs]
-
-    # Run
-    for day,source,manual_meta,saveto in zip(
-            day_root_dirs,sourceplate_files,manual_meta_files,metadata_files):
-        if len(source)!=1:
-            print('There is not a unique sourceplates file in '
-                  +'day {}. Metadata cannot be compiled'.format(day))
-            continue
-        if len(manual_meta)!=1:
-            print('There is not a unique manual_metadata file in '
-                  +'day {}. Metadata cannot be compiled'.format(day))
-            continue
-        robot_metadata = merge_robot_metadata(source[0], saveto=False)
-        day_metadata = get_day_metadata(
-                robot_metadata, manual_meta[0], saveto=saveto[0])
-
-    #%%Example 3:
-    day_root_dir = Path('/Volumes/behavgenom$/Ida/Data/Hydra/ICDbacteria/AuxiliaryFiles/20191122')
-    sourceplate_file = day_root_dir / '20191122_sourceplates.csv'
-    manual_meta_file = day_root_dir / '20191122_manual_metadata.csv'
-    metadata_file = day_root_dir.joinpath('20191122_day_metadata.csv')
-
-    plate_metadata = populate_96WPs(sourceplate_file,
-                                  entire_rows=True,
-                                  saveto= None,
-                                  del_if_exists = False)
-    day_metadata = get_day_metadata(plate_metadata, manual_meta_file, saveto=metadata_file)
-
-    #%% Example 4: syngenta screen  - use a copy of day of metadata from behavgenom
-    from tierpsytools import EXAMPLES_DIR
-
-    day_root_dir = EXMAPLES_DIR / Path('hydra_metadata/data/AuxiliaryFiles/20191213')
-    sourceplate_file = day_root_dir / '20191212_sourceplates.csv'
-    wormsorter_file = day_root_dir / '20191213_wormsorter.csv'
-    manual_meta_file = day_root_dir / '20191213_manual_metadata.csv'
-    metadata_file = day_root_dir / '20191213_day_metadata.csv'
-    bad_wells_file = day_root_dir / '20191212_robot_bad_imaging_wells.csv'
-
-    plate_metadata = populate_96WPs(wormsorter_file)
-    robot_metadata = merge_robot_metadata(sourceplate_file)
-    concat_metadata = merge_robot_wormsorter(day_root_dir,
-                                             robot_metadata,
-                                             plate_metadata,
-                                             bad_wells_file)
-    day_metadata = get_day_metadata(concat_metadata,
-                                    manual_meta_file,
-                                    saveto=metadata_file,
-                                    del_if_exists=True,
-                                    include_imgstore_name=False)
-    files_to_check = day_metadata_check(day_metadata,
-                                        day_root_dir,
-                                        plate_size=48)
-
-    # %% Example 5: map library plates to shuffled plate
-
-    sourceplate_file = Path('/Users/ibarlow/Desktop/tierpsytools_checks/SyngentaStrainScreen/2020SygentaLibrary3doses_sourceplates_run1.csv')
-
-    robot_metadata = merge_robot_metadata(sourceplate_file,
-                                          saveto=None,
-                                          del_if_exists=True,
-                                          drug_by_column=False,
-                                          compact_drug_plate=True)
-    robot_metadata.sort_values(by=['source_plate_number', 'destination_well'],
-                               ignore_index=True,
-                               inplace=True)
-
-    robot_metadata['shuffled_plate_id'] = [r.source_plate_id +
-                                           '_sh%02d' %(r.robot_run_number)
-                                           for i, r in
-                                           robot_metadata.iterrows()]
-    robot_metadata.to_csv(str(sourceplate_file).replace('.csv', '_shuffled.csv'),
-                          index=False)
-
-    # %% Example 6: with Syngenta 12 strain screen
-    from tierpsytools import EXAMPLES_DIR
-
-    day_root_dir = EXAMPLES_DIR / Path('hydra_metadata/data/AuxiliaryFiles/20200220')
-    manual_meta_file = day_root_dir / '20200220_manual_metadata.csv'
-    wormsorter_file = day_root_dir / '20200220_wormsorter.csv'
-    bad_wells_file = day_root_dir / '20200220_robot_bad_imaging_wells.csv'
-    metadata_file = day_root_dir / '20200220_day_metadata.csv'
-    sourceplates =EXAMPLES_DIR / Path('sourceplates')
-
-    sourceplates = list(sourceplates.rglob('*shuffled.csv'))
-    drug_plates = []
-    for file in sourceplates:
-        drug_plates.append(pd.read_csv(file))
-    drug_plates = pd.concat(drug_plates)
-
-    #run functions to populate
-    plate_metadata = populate_96WPs(wormsorter_file,
-                                    del_if_exists=True,
-                                    saveto='default')
-
-    bad_wells_df = convert_bad_wells_lut(bad_wells_file)
-    plate_metadata = pd.merge(plate_metadata,
-                              bad_wells_df,
-                              on=['imaging_plate_id', 'well_name'],
-                              how='outer')
-
-    print('Generating day metadata: {}'.format(
-            metadata_file))
-    try:
-        day_metadata = get_day_metadata(plate_metadata,
-                                        manual_meta_file,
-                                        saveto=metadata_file,
-                                        del_if_exists=True,
-                                        include_imgstore_name=True)
-    except ValueError:
-        print ('imgstore error')
-        day_metadata = get_day_metadata(plate_metadata,
-                                        manual_meta_file,
-                                        saveto=metadata_file,
-                                        del_if_exists=True,
-                                        include_imgstore_name=False)
-    # merge with the drug plates
-    day_metadata = pd.merge(day_metadata,
-                           drug_plates,
-                           left_on=['source_plate_id',
-                                    'well_name'],
-                           right_on=['shuffled_plate_id',
-                                     'destination_well'],
-                           suffixes=('_day', '_robot'),
-                           how='outer')
-    day_metadata.drop(
-        day_metadata[day_metadata['imaging_plate_id'].isna()].index,
-                    inplace=True)
-    # checks to see if all videos
-    files_to_check = day_metadata_check(day_metadata,
-                                        day_root_dir,
-                                        plate_size=48) # set to 48 as some half plates
-    number_wells_per_plate(day_metadata,
-                           day_root_dir)
+    pass
